+++ tests/override_config_table/test_override_config_table.py
@@ -5,13 +5,14 @@ from tests.common.utilities import skip_release
 from tests.common.utilities import update_pfcwd_default_state
 from tests.common.config_reload import config_reload
 from utilities import backup_config, restore_config, get_running_config,\
-    reload_minigraph_with_golden_config, file_exists_on_dut
+    reload_minigraph_with_golden_config, file_exists_on_dut, compare_dicts_ignore_list_order, \
+    NON_USER_CONFIG_TABLES
+
 
 GOLDEN_CONFIG = "/etc/sonic/golden_config_db.json"
 GOLDEN_CONFIG_BACKUP = "/etc/sonic/golden_config_db.json_before_override"
 CONFIG_DB = "/etc/sonic/config_db.json"
 CONFIG_DB_BACKUP = "/etc/sonic/config_db.json_before_override"
-NON_USER_CONFIG_TABLES = ["FLEX_COUNTER_TABLE", "ASIC_SENSORS"]
 
 pytestmark = [
     pytest.mark.topology('t0', 't1', 'any'),
@@ -73,21 +74,6 @@ def setup_env(duthosts, golden_config_exists_on_dut, tbinfo, enum_rand_one_per_h
 
     # Restore config before test
     config_reload(duthost)
-
-
-def compare_dicts_ignore_list_order(dict1, dict2):
-    def normalize(data):
-        if isinstance(data, list):
-            return set(data)
-        elif isinstance(data, dict):
-            return {k: normalize(v) for k, v in data.items()}
-        else:
-            return data
-
-    dict1_normalized = normalize(dict1)
-    dict2_normalized = normalize(dict2)
-
-    return dict1_normalized == dict2_normalized
 
 
 def load_minigraph_with_golden_empty_input(duthost):
