+++ tests/override_config_table/utilities.py
@@ -5,6 +5,8 @@ from tests.common.config_reload import config_reload
 
 logger = logging.getLogger(__name__)
 
+NON_USER_CONFIG_TABLES = ["FLEX_COUNTER_TABLE", "ASIC_SENSORS"]
+
 
 def backup_config(duthost, config, config_backup):
     logger.info("Backup {} to {} on {}".format(
@@ -34,3 +36,18 @@ def reload_minigraph_with_golden_config(duthost, json_data):
 
 def file_exists_on_dut(duthost, filename):
     return duthost.stat(path=filename).get('stat', {}).get('exists', False)
+
+
+def compare_dicts_ignore_list_order(dict1, dict2):
+    def normalize(data):
+        if isinstance(data, list):
+            return set(data)
+        elif isinstance(data, dict):
+            return {k: normalize(v) for k, v in data.items()}
+        else:
+            return data
+
+    dict1_normalized = normalize(dict1)
+    dict2_normalized = normalize(dict2)
+
+    return dict1_normalized == dict2_normalized
